refactor(rust): extract plugins.rs with plugin management commands
